python-devex-ai/configuration-ai-playground/functions/compose-upbox/requirements.txt
crossplane-function-sdk-python==0.5.0
pydantic==2.9.2
